==> 2_Exercise.Rmd <==
---
title: 'Exercise #2'
author: "Immanuel Frenzel"
date: "12.01.2022"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# more internal settings can go here
# Consider help pages like:
# https://rmarkdown.rstudio.com/lesson-1.html
# https://www.rstudio.com/wp-content/uploads/2015/03/rmarkdown-reference.pdf

```

### Loaded packages

```{r libraries}
library("lubridate")
library("tidyverse")
library("zoo")
library("tibbletime")
```


## 1. Quality control procedures (4 QCPs)

```{r, eval = TRUE}
#import
reimport <- read.csv("C:/Users/Imifr/Documents/Github/hyd_data_management/10610854.csv")

#time in POSIXct
data <- reimport %>% 
  mutate(dttm = ymd_hm(reimport$dttm))
```

### 1.1 Measurement range (Plausible values)

```{r qcp1, eval = TRUE}
data <- data %>%
  mutate(QCP_1 = if_else(temp <= -20 | temp >= 70, 0, 1))
```

**Question**: How many data points are outside the measurement range?

```{r, eval = TRUE}
summarise(data, QCP_1 = sum(QCP_1 == 0, na.rm = TRUE))
summary(data)
```

**Answer**: 0, all 317 Values are NAs


### 1.2 Plausible rate of change

```{r qcp2}
data <- data %>%
  mutate(QCP_2 = ifelse(between(temp - lag(temp), -1, 1), 1, 0))
```

**Question**: Describe shortly how many data points failed during this QCP and discuss whether there is a certain daytime pattern of failure or not?

**Answer**: 7 points failed QCP_2. All readings are between 08:20:00	and 11:30:00. This makes sense. Qick temperature rise is expected for morning times.

```{r, eval = TRUE}
summarise(data, QCP_2 = sum(QCP_2 == 0, na.rm = TRUE))
filter(data, QCP_2 == 0)
summary(data)
```

### 1.3 Minimum variability (Persistence)

```{r qcp3}
data <- data %>%
  mutate(QCP_3 = ifelse((temp == lag(temp)) +
                        (temp == lag(temp, n = 2)) +
                        (temp == lag(temp, n = 3)) +
                        (temp == lag(temp, n = 4)) +
                        (temp == lag(temp, n = 5)) == 5, 0, 1 
                        ))
```

**Task**: Code in this section should analyses the persistance.

```{r}
filter(data, QCP_3 == 0)
```



### 1.4 Light intensity

```{r qcp4}
data <- data %>%
  mutate(SIC = case_when(lux < 0 ~ "NA",
                         lux < 10 ~ "night",
                         lux < 500 ~ "sun_rise_or_set",
                         lux < 2000 ~ "overcast_full",
                         lux < 15000 ~ "overcast_light",
                         lux < 20000 ~ "clear_sky_shady",
                         lux < 50000 ~ "sunshine",
                         lux >= 50000 ~ "sunshine_bright")) %>% 
  mutate(QCP_4 = case_when(is.na(SIC) ~ as.double(NA),
                           hour(dttm) < 6 | hour(dttm) >= 18 ~ 1, #set all QCP_4 values for nighttime to 1
                           lag(SIC, n = 3) == "sunshine_bright" ~ 0, #set 0 if sunshine_bright +- 3 before or ahead
                           lag(SIC, n = 2) == "sunshine_bright" ~ 0,
                           lag(SIC) == "sunshine_bright" ~ 0,
                           SIC == "sunshine_bright" ~ 0,
                           lead(SIC) == "sunshine_bright" ~ 0,
                           lead(SIC, n = 2) == "sunshine_bright" ~ 0,
                           lead(SIC, n = 3) == "sunshine_bright" ~ 0, #set 0 if sunshine +- 1 before or ahead
                           lag(SIC) == "sunshine" ~ 0,
                           SIC == "sunshine" ~ 0,
                           lead(SIC) == "sunshine" ~ 0,
                           TRUE ~ 1)) 
```

**Task**: Discuss shortly how often and when during daytime the QCP4 flags bad data. Elaborate on some reasons for your results.

```{r}
summary(data)
ggplot(data, mapping = aes(x = SIC)) +
  geom_bar() +
  theme_bw()
```

**Answer**: 0,  It seems like there was not much light at the sensor location.

## 2. Synthesis

```{r synthesis}
a <- data %>% 
  summarise_at(vars(QCP_1 , QCP_2, QCP_3, QCP_4), ~ sum(.x == 0, na.rm = TRUE)) %>% 
  pivot_longer(cols = QCP_1:QCP_4)

ggplot(a, mapping = aes(x = name, y = value)) +
  geom_col() +
  theme_bw()
```

**Task**: Present a table or graph to show how many data points fail during the four specific QCPs. Discuss shortly the reasons for failure and compare the different QCPs against each other.

**Answer**:  Reasons for failure were sudden temperature changes (QCP_2, 7 points) and constant temperature over at least one hour (QCP_3, 45 points). All temperature readings were in the measurement-interval (QCP_1). The sensor never experienced more than 5166.7 lux (QCP_4) which was not enough for not passing the checkpoint.

## 3. Results

### 3.1 Result (Flagging system: 10-minutes-values)

```{r res1}
qc_df <- data %>% 
  mutate(QCP_total = ifelse(QCP_1 + QCP_2 + QCP_3 + QCP_4 < 4, 0, 1))

head(qc_df)
```

**Task**: At the end of the code section above you should generate one! tibble or data.frame named `qc_df` with all time information, all data points (temperature and lux) and your outcomes of the different QCPs.

### 3.2 Result (Aggregate to hourly series)

```{r res2}
hobo_hourly <- qc_df %>% 
  mutate(hour = cut(dttm, breaks = "hour")) %>% 
  group_by(hour) %>% 
  summarise(date_time = first(hour), th = round(ifelse(sum(QCP_total) < 6, NA, mean(temp)), digits = 4)) %>% 
  select("date_time", "th")

head(hobo_hourly)
```

**Task**: At the end of the code section above you should generate one! tibble or data.frame named `hobo_hourly` with averaged temperature values per hour or NA values (if the hour is flagged as bad data). See exercise description for more details.

-   First column: YYYY-DD-MM HH:MM:SS

-   Second column: Temperature values (4 digits), NA values possible
