==> 3_Exercise.Rmd <==
---
title: "Exercise 3"
author: "Immanuel Frenzel"
date: "13 1 2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

### Loaded packages

```{r libraries, echo = FALSE}
library("lubridate")
library("tidyverse")
library("zoo")
library("tibbletime")
```

```{r cleanspace, echo = FALSE}
rm(list=ls())
```

```{r import}
my_hobo <- read.csv("https://raw.githubusercontent.com/imifrenzel/hyd_data_management/main/10610854_hourly.csv")
my_hobo <- my_hobo %>% 
  mutate(dttm = ymd_hms(date_time)) %>% 
  mutate(my_hobo = th) %>% 
  select("dttm", "my_hobo")

dwd_airport <- read.csv("https://raw.githubusercontent.com/jasperschalla/station_data/master/dwd_airport.csv")
dwd_airport <- dwd_airport %>% 
  mutate(dttm = ymd_hms(dttm))

dwd_urban <- read.csv("https://raw.githubusercontent.com/jasperschalla/station_data/master/dwd_urban.csv")
dwd_urban <- dwd_urban %>% 
  mutate(dttm = ymd_hms(dttm))

uni_meteo <- read.csv("https://raw.githubusercontent.com/jasperschalla/station_data/master/uni_meteo.csv")
uni_meteo <- uni_meteo %>% 
  mutate(dttm = ymd_hms(dttm))

wbi <- read.csv("https://raw.githubusercontent.com/jasperschalla/station_data/master/wbi.csv")
wbi <- wbi %>% 
  mutate(dttm = ymd_hms(dttm))
```

```{r join}
data <- left_join(my_hobo, dwd_airport, by = "dttm") %>% 
  left_join(., dwd_urban, by = "dttm") %>% 
  left_join(., uni_meteo, by = "dttm") %>% 
  left_join(., wbi, by = "dttm")

```

```{r long format}
data_long <- data %>%
pivot_longer(cols = -c(dttm), 
names_to = "station",
values_to = "temp")

head(data_long)
```

```{r plot}
ggplot(data = data_long, aes(x = dttm, y = temp)) +
  geom_line(aes(lty = station, colour = station)) +
  theme_bw() +
  labs(title = "Comparison across stations",
    x = "Date", y = "Temperature (°C)")

```

## DWD_Urban matches best
```{r linear models}
summary(lm(my_hobo ~ dwd_airport, data))$r.squared
summary(lm(my_hobo ~ dwd_urban, data))$r.squared
summary(lm(my_hobo ~ uni_meteo, data))$r.squared
summary(lm(my_hobo ~ wbi, data))$r.squared
```
Seems like dwd_urban fits best

```{r linearregression with dwd_urban}
export <- data %>% 
  mutate(th = ifelse(is.na(my_hobo), round(0.812342*dwd_urban+3.562775, 3), round(my_hobo, 3))) %>% 
  mutate(origin = ifelse(is.na(my_hobo), "R", "H")) %>% 
  select("dttm", "th", "origin")
```

## Export
```{r}
write.csv(export, file = "C:/Users/Imifr/Documents/Github/hyd_data_management/10610854_Th.csv", append = FALSE, quote = FALSE, sep = ",",
          eol = "\n", na = "NA", dec = ".", row.names = FALSE,
          col.names = TRUE, qmethod = c("escape", "double"),
          fileEncoding = "")
```
