==> 4_Exercise.Rmd <==
---
title: "4_Exercise"
author: "Immanuel Frenzel"
date: "14 1 2022"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r library and clean workspace}
library("tidyverse")
library("lubridate")
library("tibbletime")

rm(list=ls())
```
## import
```{r import, echo = F}
data <- read.csv("https://raw.githubusercontent.com/imifrenzel/hyd_data_management/main/10610854_Th.csv")
data <- data %>% 
  mutate(dttm = ymd_hms(dttm)) %>% 
  tibble()

data_ex3 <- read.csv("https://raw.githubusercontent.com/imifrenzel/hyd_data_management/main/10610854_hourly.csv")
data_ex3 <- data_ex3 %>% 
  mutate(dttm = ymd_hms(date_time)) %>% 
  tibble()

data_ex1 <- read.csv("https://raw.githubusercontent.com/imifrenzel/hyd_data_management/main/10610854.csv")
data_ex1 <- data_ex1 %>% 
  mutate(dttm = ymd_hm(dttm)) %>% 
  tibble()
```

## tnight tday tavg
```{r tnight tday tavg, echo=TRUE}
tdaynight <- data %>% 
  mutate(daytime = ifelse(hour(dttm) < 6 | hour(dttm) >= 18, "night", "day")) %>% 
  group_by(daytime) %>% 
  summarise(mean = mean(temp))

tavg <- mean(data$temp)

head(tdaynight)
tavg
```
## tamp
```{r tamp}
tamp_data <- data %>%
  mutate(date = date(dttm)) %>% 
  group_by(date) %>%
  summarise(tmin = min(temp), tmax =(max(temp)))

tamp = mean(tamp_data$tmax) - mean(tamp_data$tmin)
tamp
```
## t6h
```{r t6h}
t6h_data <- data %>%
  mutate(t6h_1 = abs(temp - lead(temp)),
         t6h_2 = abs(temp - lead(temp, n = 2)),
         t6h_3 = abs(temp - lead(temp, n = 3)),
         t6h_4 = abs(temp - lead(temp, n = 4)),
         t6h_5 = abs(temp - lead(temp, n = 5)),
         t6h_6 = abs(temp - lead(temp, n = 6))) %>%
  rowwise() %>%
  mutate(t6h = max(t6h_1, t6h_2, t6h_3, t6h_4, t6h_5, t6h_6, na.rm = T))

head(t6h_data)
tail(t6h_data)

t6h <- max(t6h_data$t6h)
t6h
```
Last value for t6h_data$t6h is -Inf due to teh algorythm, this should be no problem for futher calculation

## lavg
```{r lavg}
data_lavg <- data_ex1 %>% 
  mutate(daytime = if_else(hour(dttm) < 6 | hour(dttm) >= 18, "night", "day")) %>% 
  group_by(daytime) 

head(data_lavg)

median_lux <- summarise(data_lavg, median = median(lux, na.rm = T))
median_lux
```

## lmax
```{r lmax}
data_lmax <- data_lavg %>% 
  ungroup() %>%
  filter(daytime == "day") %>% 
  mutate(hm = hm(format(dttm, "%H:%M"))) %>% 
  group_by(hm) %>% 
  summarise(meanlux = mean(lux, na.rm = TRUE)) %>% 
  arrange(desc(meanlux))

head(data_lmax)

lmax <- data_lmax[1,1]
lmax
```
